requirements.txt
asgiref==3.5.2
Django==4.0.4
django-browser-reload==1.12.0
django-tailwind==3.6.0
gunicorn==20.1.0
psycopg2==2.8.6
sqlparse==0.4.2
whitenoise==6.2.0
python-dotenv==1.0.0
django-phonenumber-field==7.2.0